+++ src/traits.rs
@@ -2,12 +2,10 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-#[cfg(feature = "parsing")]
 use svgparser::TextFrame;
 
 use WriteOptions;
 
-#[cfg(feature = "parsing")]
 /// The trait for parsing data from the data stream.
 pub trait FromStream: Sized {
     /// Error type.