+++ src/types/path/parser.rs
@@ -2,16 +2,13 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-#[cfg(feature = "parsing")]
 use svgparser;
-#[cfg(feature = "parsing")]
 use svgparser::{
     Error as ParseError,
     TextFrame,
     Tokenize,
 };
 
-#[cfg(feature = "parsing")]
 use FromStream;
 
 use super::{
@@ -20,7 +17,6 @@ use super::{
     SegmentData,
 };
 
-#[cfg(feature = "parsing")]
 impl FromStream for Path {
     type Err = ParseError;